Deduplicate & move saveRaw to base class
