+++ ghost/core/core/server/adapters/storage/LocalMediaStorage.js
@@ -1,11 +1,7 @@
 // # Local File System Media Storage module
 // The (default) module for storing media, using the local file system
-const fs = require('fs-extra');
-const path = require('path');
 const config = require('../../../shared/config');
-
 const constants = require('@tryghost/constants');
-const urlUtils = require('../../../shared/url-utils');
 const LocalStorageBase = require('./LocalStorageBase');
 
 const messages = {
@@ -22,29 +18,6 @@ class LocalMediaStorage extends LocalStorageBase {
             siteUrl: config.getSiteUrl(),
             errorMessages: messages
         });
-    }
-
-    /**
-     * Saves a buffer in the targetPath
-     * @param {Buffer} buffer is an instance of Buffer
-     * @param {String} targetPath relative path NOT including storage path to which the buffer should be written
-     * @returns {Promise<String>} a URL to retrieve the data
-     */
-    async saveRaw(buffer, targetPath) {
-        const storagePath = path.join(this.storagePath, targetPath);
-        const targetDir = path.dirname(storagePath);
-
-        await fs.mkdirs(targetDir);
-        await fs.writeFile(storagePath, buffer);
-
-        // For local file system storage can use relative path so add a slash
-        const fullUrl = (
-            urlUtils.urlJoin('/', urlUtils.getSubdir(),
-                this.staticFileURLPrefix,
-                targetPath)
-        ).replace(new RegExp(`\\${path.sep}`, 'g'), '/');
-
-        return fullUrl;
     }
 }
